Adopts Retrievable Explanation (#250)
## What is the goal of this PR?
As of Grakn 1.6.0 and since the changes made in the explanation retrieval with Client Node.js, explaining the inferred concepts were not possible against Core 1.6.0. This PR adopts these changes and brings back the explanation feature of Workbase.

## What are the changes implemented in this PR?

### Attaching `queryPattern` to nodes
Within the `EXPLAIN_CONCEPT` action in order to know if the retrieved explanation is a joint explanation or one resolved from a rule, we check the variables of the explanation answers against the original `queryPattern`. For this reason, we need to have access to the node's `queryPattern` and so we attach it to the node when building the node object.

### Retrieving explanation
In the legacy implementation of explanation retrieval, we could safely assume that when `queryPattern` is an empty string, the explanation is of type Joint and so we need to retrieve explanations one level deeper. This is no longer the case, and therefore we rely on the `isJointExplanation` variable which checks the variables of the retrieved explanation answers against the original `queryPattern`.