Improves performance of retrieving attributes and role players (#282)
## What is the goal of this PR?
Changes introduced in this PR are aimed at improving the performance of Workbase, by using the Concept API instead of constructing and running native Graql queries.

## What are the changes implemented in this PR?

### Attributed Retrieval
The post-processing step that occurs after visualising the graph for the user query, now uses the Concept API (instead of constructing and running a Graql query) to fetch the attributes of each node. Although this results in much faster retrieval of attributes, it does come with the limitation of missing inferred attributes. we will revert this change, once we have optimised the query planner for fast retrieval of attributes (of a given instance) through a Graql query.

### Fetching roleplayers (_Load Roleplayers_ enabled)
When the user enables the _Load Roleplayers_ option and queries for a relation, without specifying its roleplayers, prior to this PR Workbase would run a Graql query for each relation instance in order to fetch its roleplayers and visualise them. This now happens via the Concept API which results in a significant improvement in the query time.