Fixes a bug where `getRule` gets run on a joint explanation (#300)
## What is the goal of this PR?
Explanation now works as expected when the first explanation is joint. This marks the 1.3.1 release.

## What are the changes implemented in this PR?
While introducing this fix I came to realise that the _unassigned variables_ error message will not be displayed when the target explanation is within the answers of the joint explanation. I have created an issue for it with high priority: https://github.com/graknlabs/workbase/issues/299