Visualiser actions: filter out explanation answers that are not part of the node's explanation (#315)
## What is the goal of this PR?
Given that all inferred concepts retrieved within a single answer share the same explanation object, we now ensure that when explaining a concept, only the concepts that are a part of the specific explanation of the given concept are visualised, as opposed to concepts that explain the other inferred concepts of the same explanation object. 

## What are the changes implemented in this PR?
When retrieving the rule explanations of answers (of a join explanation), we now filter out the answers that are unrelated to the node that's being explained.