Improves presentation and logic of attributes displayed on a node's label (#328)
## What is the goal of this PR?
When selecting attribute labels in the Display settings for a given type, the attribute label is added to the nodes' label even though the instance (i.e. node) does not own an instance of the selected attribute type. In this case, the attribute's value is displayed as empty. In addition, commas are no longer displayed at the end of each attribute label/value on the node's label. 

## What are the changes implemented in this PR?
resolves https://github.com/graknlabs/workbase/issues/329
resolves https://github.com/graknlabs/workbase/issues/325