add AppImage to linux targets (#334)
## What is the goal of this PR?

It has been really painful for me to try and get the tar.gz build to work.
After some struggle I managed to get the app to launch but the screen would be blank.
Building .AppImage and executing that worked like a charm. It launched, there was no blank screen and the UI was usable.

Benefits of using AppImage: https://docs.appimage.org/introduction/advantages.html#advantages-for-users

## What are the changes implemented in this PR?

Add AppImage as a target for Linux.