Upgrade workbase to 2.0 (#363)
## What is the goal of this PR?

We upgraded Workbase to support Grakn 2.0. We can now perform schema and data queries against a Grakn 2.0 server (either Core or Cluster) and visualise the results and their neighbours.

Note that this is a **breaking change** and the software no longer supports the 1.x versions of Grakn. For legacy versions of Grakn, the previous releases of Workbase should be used.

## What are the changes implemented in this PR?

Upgrade Workbase to Grakn 2.0