TypeDB Studio Alpha (#454)
## What is the goal of this PR?

We rebuilt TypeDB Workbase in order to allow us to more easily add new, cool features to the product. The original was built in Electron and React, and we found it did not scale well with the complexity of the application. Additionally, it suffered from a severe performance issue which rendered it frustrating to use.

The new TypeDB Studio will feel much more like a full-blown Integrated Development Environment, supporting loading and saving TQL files and workspaces, outputting results not just to a graph but also to a Log and a tabular view, enabling the processing of all kinds of queries - even aggregates and inserts. These features will come in future releases, but soon! It is built using Kotlin and JetBrains Compose, and run on a JVM which is bundled into the installer.

For now, we support:

- basic Match functionality;
- reasoning and explanations

Currently the main improvement in Studio is performance. The time taken to execute Match queries is minimal.

## What are the changes implemented in this PR?

Rewrite TypeDB Workbase as Studio, using Kotlin and Compose