Refactor logic for opening and activating Resources
Concentrated all logic opening and activating Resources into each resource implementation class: File and TypeState.Thing. This allows us to minimise cycles between each type of Resource and the ResourceManager, as well as create a consistent API that forces users to always open a Resource by calling .open() and .activate() on the Resource object directly, which in turn will register itself to the ResourceManager.