Fix possible error when drawing an attribute's edge (#585)
We bring the function for finding an attribute's edge endpoint in line with relation and entity's functions by allowing the offset to be nullable.

The user should see fewer crashes in the unlikely case that their schema causes the graph rendering logic to attempt to divide by zero.

We change the function signature of `ellipseIncomingLineIntersect` and `edgeEndpoint` to allow `Offset` to be nullable and detect this case in the body of `ellipseIncomingLineIntersect`.