Expand vertices on hover/focus, and wrap content by logical block (#590)
## What is the goal of this PR?

Vertices will now expand on hover/focus. In the case of hover, this replaces the old "fade" effect.

When the content of the vertex overflows the base size, we expand the vertex significantly more, inflating it so it can show much more content - e.g: the full label of a relation, or the full value of an attribute. (Exceptionally long values may still be truncated, but the full value is retrievable in the Preview pane.)

Also, vertex content is now wrapped by logical block. So, for instance, type labels will break onto the next line at the positions of hyphens, rather than whenever they overflow the vertex bounds.

## What are the changes implemented in this PR?

- Don't show root types in Graph Visualiser
- Expand vertices on hover/focus, replacing the old "fade" animation on hover
- Divide vertex labels by logical block, not just by overflow