Test suite for Studio (#605)
## What is the goal of this PR?

In the course of fulfilling https://github.com/vaticle/typedb-studio/issues/392 we have developed a suite of tests concerning the integration of discrete sections of Studio. 

Automated testing brings benefits to the future development of Studio: We can be sure before releasing new versions that changes have not broken previous functionality without anyone having to test each of the components and confirm that they work, giving users more confidence in the application.

Developers can now easily add tests to the suite when fixing bugs so that regressions in functionality and error-free use do not arise between releases.

## What are the changes implemented in this PR?

We introduce tests for the Project Browser, the Text Editor, the Type Browser and update Quickstart to make use of the interactive methods of writing schemas and data that we've identified during the course of development.

In the course of testing we sometimes manipulate the state directly to bridge small gaps that are left by incomplete parts in the provided test framework from compose-jb. This is seen in our addition of `//test/integration:__pkg__` to various BUILD files. It has been agreed that this is acceptable as long as it is the minimum necessary to test.