Schema Manager now handles read transactions being null
Prior to this change, it is possible for the `SchemaManager.openOrGetReadTx()`
to return a null TypeDBTransaction, which is not expected (as declared by the
method signature). The cause of this issue was the following code:

```
private var readTx: AtomicReference<TypeDBTransaction> = AtomicReference()
```

Even though `AtomicReference` is declared to not take nullable `TypeDBTransaction`,
Java allows you to pass `readTx.set(null)` which would be illegal if
`AtomicReference` is a Kotlin class. But it's not. It's a Java class. The
issue becomes worse when `readTx.get()` is not recognised to return nullable,
because the generic signature was not nullable!

We have now declared `readTx` to be `AtomicReference<TypeDBTransaction?>` and
`openOrGetReadTx()` to return `TypeDBTransaction?` (nullable), and propagated
the change to all callers of this method to handle nullable transaction objects.