QueryRunner records last response time correctly for response streams
Prior to this change, the QueryRunner had a bug, where responses in a
stream did not seem to affect the "last response time" value that affects
when the running indicator ("...") should be printed. The running indicator
was supposed to be printed every 3 seconds if there is no incoming response
from the server. This bug caused QueryRunner to think that there were no
response from the server when there actually is -- the stream responses.

We fixed this issue by making sure we update the last response time every
time we consume a response in the stream.

This commit closes #617.