Refactored LogOutput output function construction
LogOutput works by providing methods that build "output functions" for RunOutputGroup output consumer to invoke. In this refactor, we have streamlined these outputFn() methods to all have a consistent and uniform signature, and ensured that they all in the end call the same private method output(type, text) in LogOutput. This refactor is a precursor to fixing issue #618.