Migrated logic to print running query indicator into LogOutput
When running a query that is slow to print out the next result, the log output is meant to print '...' to indicate to the user that the application is still running and currently waiting for a response from the server. However, this functionality broke when we rearchitected the concurrency model of RunOutputGroup for optimal performance -- all the '...' indicator was printed in bulk at the end of the log output, which defeats its purpose. This happened because the QueryRunner response collection model was misaligned to how it was consumed by the RunOutputGroup.

We have now fixed this problem by migrating the logic to print running query indicator into LogOutput directly, such that QueryRunner or RunOutputGroup no longer needs to care about managing the indicator for LogOutput. LogOutput now keeps track of when the last output was printed, and if 3 seconds has lapsed since the last output, it will print '...'.

This commit fixes #618.