Closing a query runner output tab confirms that the query is not running
Prior to this change, users are able to close the query runner output tab on a query that is currently running. This breaks the UX pattern that we have, where pages can't be closed until their queries are stopped from running. The motivation is not leave any dangling queries running in the background that users are aware and in control of.

This change now enforces the user to have stopped the running query associated to the runner output tab.

This issue fixes #616.