Exceptions thrown while consuming answers from TypeDB are now handled
Prior to this change, we had a bug, where an exception that occurs during consuming answers from TypeDB, such as transactions being closed for whatever reason (could also be user initiated) would not be handled properly, causing QueryRunner to end the response stream properly with a 'Done' message. This lead to RunOutputGroup to continue running, waiting for QueryRunner to send a 'Done' message, indefinitely.

In this change, we now handle any exception that may occur during the consumption of answers from TypeDB Server, by logging the error, and ensuring we always conclude the response stream with a 'Done' message.

This issue fixes #619.