Enable renaming project directory from inside and outside Studio
Prior to this change, renaming the project directory from ProjectBrowser would crash the Studio application. Same would be true if we rename the directory from outside Studio. This change now enables us to rename the project directory from within the application, through the ProjectBrowser. This change also allows the application to monitor the filesystem for changes to the project directory, and handles the error caused by the the project directory going being changed or becoming inaccessible.