Fix window context not reflecting resizing (#621)
## What is the goal of this PR?

We fixed a bug in `WindowContext` where resizing the underlying window was not reflected when getting its values (i.e. height, width, x and y). This fixes a bug where certain actions such as mouse hovering didn't render tooltips because of windowContext reporting the incorrect height and width.

## What are the changes implemented in this PR?

We've expanded WindowContext so that it includes 2 inner classes:
- Test, a 'mock' testing window context.
- Compose, a window context that links to the window itself when Studio is launched with a window.