Migrate and stabilise Studio test suite (#620)
## What is the goal of this PR?

The Studio tests that required an instance of TypeDB have been migrated to using `typedb_kt_test`, which use `typedb-runner` to provide TypeDB to the tests rather than `typedb-extractor`. This migration has provided stability, increased hermeticity and the opportunity to test Studio's interactions with cluster in the future.

## What are the changes implemented in this PR?
Studio tests that connect to TypeDB now use `typedb_kt_test`, a newly provided test rule from typedb-common that allow us to make use of TypeDB runners.

Utils is now an Object rather than a standalone set of functions and variables. It also provides a new test function `studioTestWithRunner`. Some Studio tests don't require the use of TypeDB and waiting for TypeDB to run is wasteful in those instances.

Studio tests are now stable in CI through an increase in delays to allow for the less compute-heavy machines in CI time to process UI recomposition and perform local network actions.