Introduce write transaction in SchemaManager
Prior to this change, SchemaManager only provides a read transaction to display information in the TypeBrowser and TypePage. To prepare for the implementation of schema editing functionalities in the TypePage, we now have to provide a write transaction through SchemaManager. However, this write transaction needs to also be served as the read transaction of SchemaManager, as all read operations should go through the same write transaction once a write transaction has been opened for editing the schema on the database. This is now what is implemented through this commit.