Fixed logic and usage of loading of TypeState properties
Prior to this change, it's possible for the TypeState to recursively call loadProperties(), asynchronously. Additionally, the method loadSubtypesExplicit() was not thread-safe and resulted in incorrect logic when access concurrently. The combination of both issues resulted in TypeState loading incorrect (duplicate) subtypes. This has now been fixed by making sure loadProperties() is called only when necessary (not redundantly and incorrectly recursive) as well as making loadProperties() synchronized.