SchemaManager refreshes & prune types when write tx closes
Prior to this change, it is possible for the user to create a type in the schema and open the type page, before committing. If the transaction is closed without committing, the page will remain opened, incorrectly. This has now been fixed by implementing a new behavior in SchemaManager, where we refresh and prune the types loaded in the SchemaManager after a schema write transaction closes, which includes committing.