Allow creating entity/relation/attribute subtypes interactively
Prior to this commit, we are not able to make any changes to the schema interactively through the TypeBrowser or TypePage. We can only modify the schema by running define queries via the query runner. With this change we have now implemented the basic foundations of schema (type) editing functionalities, to be performed interactively through the TypeBrowser and TypePage. This change also already includes the ability to create entity, relation, and attribute types and subtypes of pre-existing user defined types. We have laid down the ground work for further editing of a type, such as renaming, editing abstract, and editing supertype.