Purge uncommitted types properly when closing SchemaManager
Prior to this change, closing a SchemaManager that had uncommitted types is problematic. If we close a SchemaManager that had uncommitted types, they did not get pruned properly. This results in an error when the user reopens a session to the same database and opens the CreateTypeDialog, as the dropdown tries to refresh the list of types.

Additionally, we uncovered a bug in SessionState where a TransactionState can be created even when SessionState is already closed. The same fix was applied to TransactionState creation of QueryRunner.