Refactored TypeState to use generics and allow changing supertypes
Prior this commit, TypeState was not implemented in such a way that it capture generic logic across its subclasses: Entity, Attribute, Relation and Role. This created quite a bit of duplicated code. We have now cleaned this up by generalizing TypeState with the necessary class parameter generics, and encapsulate all repeated logic efficiently in TypeState.

As the issue above was discovered while we were implementing the feature of allowing users to change the supertype of a given type, this commit additionally contains changes to implement this new feature too.