Use icons by purpose rather than shape, use consistent casing in enums (#629)
## What is the goal of this PR?

We address icons by their purpose rather than by their shape. This leads
to a more ergonomic development story where the icon is named by how we
use it. For example, the icon for exporting a schema is now
`Icon.Purpose.EXPORT` rather than
`Icon.Code.ARROW_UP_RIGHT_FROM_SQUARE`.

Additionally, we want to have consistent casing for enum members.

## What are the changes implemented in this PR?

Previously, we addressed functions by their shape through `Icon.Code`.
However, this lead to unergonomic scenarios where we wanted an icon but
addressed it by a vague description given to us by our icon pack rather
than how we use it within the interface.

We have introduced a new enum class named `Purpose`, through which we
alias the icons we have imported to their actual purpose rather than
their shape/a context-less description. Additionally, we have renamed
`Code` to `Shape`, and describe the icons in `Shape` by their shape and
not a mix of their purpose and their shape as we have done previously.

Additionally, we refactored out use of PascalCase for enum members in
favour of SCREAMING_SNAKE_CASE as it is the standard for Kotlin.