Toggling infer no longer toggles explanations (#633)
## What is the goal of this PR?

When we run a query, we automatically include `TypeDBOptions` given the
state of the toolbar. Currently, we derive the value of `explain` for
`TypeDBOptions` from `infer`. We should derive it from `explain`.

## What are the changes implemented in this PR?

We only turn on explanations when explain is on, not when `infer` is on.

We have also added a regression test to ensure that this issue does not
return.