Add preferences dialog (#623)
## What is the goal of this PR?

We have added a preferences dialog that is accessible by a settings
(gear) icon in the top right corner of Studio. This dialog will allow
users to make small adjustments to various functions of Studio.
Developers will be more able to add preferences easily in the future
given this framework.

This PR closes the following issues:
- https://github.com/vaticle/typedb-studio/issues/607
- https://github.com/vaticle/typedb-studio/issues/435
- https://github.com/vaticle/typedb-studio/issues/484
- https://github.com/vaticle/typedb-studio/issues/550

and will enable us to work on:

- https://github.com/vaticle/typedb-studio/issues/428
- https://github.com/vaticle/typedb-studio/issues/463
- https://github.com/vaticle/typedb-studio/issues/569
- https://github.com/vaticle/typedb-studio/issues/573
- https://github.com/vaticle/typedb-studio/issues/600

## What are the changes implemented in this PR?

- We've added `PreferenceDialog`, an interface for users to change their
preferences for how Studio behaves. `PreferenceDialog` is responsible for
the interface through which people change their preferences.
- We've extended `PreferenceManager` to provide an interface for the rest
of the application to get preference data from.
- We've added a inner Preferences class to DataManager so that we can
save our preferences in the properties file we already store.
- We've added `PreferenceField`, which are grouped by `PreferenceGroup`s.
- Dialog's Layout allows for the opting out of padding. We do this to
achieve the desired look and feel for `PreferenceDialog`.
- iconArgs are optional for Navigators. We do this so that preferences
don't need to have associated icons to their left.
- We've sorted Label.
- We've implemented 4 preferences: Auto saving for the text editor,
making graph output optional, ignored paths and limiting match query
results.

## Additional notes

To any reviewer - please check out this branch and use the dialog
yourself! The look and feel is important to get right, so please comment
if you have any thoughts.