Fix replace shortcut for text editor (#640)
## What is the goal of this PR?

The shortcut to open the replace dialog in Studio **on macOS** was
`CTRL+R`, but it should be `CMD+R`. This has now been fixed.

## What are the changes implemented in this PR?

This is a very small change 1+ 1- change, moving the line from under the
CTRL modifier case to the CMD modifier case. Users should now be able to
run `CMD+R` to open the replace functionality in Studio.