Improve preference dialog canonicalness (#643)
## What is the goal of this PR?

We have improved the code providing the preferences dialog so that it is
easier to understand and maintain in the future. The end user should
experience no impact. Contributors and developers should have an easier
time reading, understanding, maintaining and extending and interacting
with the preferences dialog in the future.

## What are the changes implemented in this PR?

We have 
- eliminated unnecessary uses of `val` in constructors where we don't
intend to create a property for a class and only require the argument
for instantiation.
- converted `PreferenceField` into a class from an interface, since it
provides implementation and data generic across its descendent classes.
- moved functionality that is the same across descendent classes into
the parent `PreferenceField` class.
- disambiguated the use of `Checkbox`, `TextInputField` and `Dropdown`,
e.g. using `Form.Checkbox` instead of `Checkbox` as another `Checkbox`
exists within the scope of `PreferenceDialog`.
- made private values `private`.