Improve certificate selection for cluster connections (#641)
## What is the goal of this PR?

We provide users connecting to cluster servers the ablility to browse
their file system and select cluster certificates, rather than having to
type the filepath out (or copy paste it from their file system browser).

As a result of this PR, end users will an easier time selecting
certificates for cluster server connections in the future. Contributors
and developers have an abstracted file selector dialog interface that
they may re-use throughout Studio.

## What are the changes implemented in this PR?

The FileSelectorDialog object has been introduced, providing functions
to Studio to present users with a dialog to select files or folders.

The CACertificateFormField in ServerDialog has been updated to provide a
button for selecting the CA certificate.