Fix icon clipping (#664)
## What is the goal of this PR?

We've fix a bug we introduced a bug in:
- https://github.com/vaticle/typedb-studio/pull/629

where the size of icons was initialised as "Unspecified". As a result,
the icons became 'clipped'.

We believe the cause of this is that the function `.sp` doesn't have
enough information in order to construct a `TextUnit` value at the time
that the object is constructed. As such, it gives a `TextUnit` value of
"Unspecified".

## What are the changes implemented in this PR?

The only other place we use `.sp` in our application is in
`//framework/common/theme/Typography.kt`, where we follow the pattern of
declaring our sizes as `const` Ints and transforming them into
`TextUnit` values later with `.sp`. We now follow this pattern in
`//framework/material/Icon.kt`

We've also made all the Icon enum class constructor arguments named.