Code and architecture review for release 2.14.1 (part 2)
7. `//images` should be under `//docs`

Since the content of `//images` are purely to be used for `README.md` -- which is "documentation", we should put `//images` under `//docs`. What may have you put you off from putting them under `//docs` in the first place was the fact that `//docs` contained "package structure diagrams". This is not an unusual situation. What you should do is refactor `//docs` such that all of its contents move into a sub-package. In this case, we introduce a sub-package `//docs/packages`. Now we can move `//images` to `//docs/images`.

8. File names under `//images` should use kebab-case

I know we discussed in the team that we want to migrate to using snake_case instead of kebab-case a while ago. However, we have not applied this change in TypeDB Studio repo, and non source code files right now are still named using kebab-case. We should maintain consistency in the codebase, and migrate all file names at once when we decide to switch.