Code and architecture review for release 2.14.1 (part 1)
1. `WindowContext` properties' visibility and setter methods

Properties of `WindowContext` are never actually modified, only read, and thus should only be `val` instead of `var`. But not only that, the properties of `WindowContext.Compose` are actually bugs as they stand right now: every property had a setter but each setter function did nothing. When the next developer uses this module and try to update the value of each property, and they will not realise that that value does actually get updated. And actually, IntelliJ highlighter already gave a strong code smell away: IntelliJ even gave a yellow warning sign on each line where the value passed onto the setter function was not assigned to anything.

2. Method naming of `InputTarget.moveCursorToStart()` and `InputTarget.moveCursorToEnd()`

The renaming of `InputTarget.moveCursorToStart()` and `InputTarget.moveCursorToEnd()` to both be `moveCursorToStartOfFile()` and `moveCursorToStartOfFile()`, is not actually accurate. You're letting the knowledge of "how the function is used externally" bleed into your naming of strictly "what the function does". Let me give you a scenario where this breaks: the TextEditor can be used to display and modify text that is not just from a file. For example, the TextEditor is used to view the log output of query runners -- the TextEditor is set to run in "read-only" mode. You see how start/end "of file" does not make sense here cause there is no file? When naming a function (or any module code), you need to make sure you're assigning a generic name that describes strictly what it does, agnostic/independent of knowledge of the users of the function. In this case, I don't think `start` and `end` is ambiguous in the context of what the TextEditor is able to manipulate, which is the "text content". But if you need do clarify it further, it should be using terminology that is already used in this module, which would be "content". Thus, `moveCursorToStartOfContent()` and `moveCursorToEndOfContent()`.

3. Changing the events for `SELECT_HOME` and `SELECT_END`

Why did we change the event for `SELECT_HOME` and `SELECT_END` to be `moveCursorToStartOfLine()` and `moveCursorToEndOfLine()`? I can't test this on the plane, as I only have laptop keyboard which doesn't have home & end keys, but as you identified yourself above, `MOVE_HOME` and `MOVE_END` means the cursor jumps to the start/end of the file/content. And given that every `SELECT_X` event is always equivalent to the `MOVE_X` event, but with `isSelecting=true`, doesn't it mean that `SELECT_HOME` and `SELECT_END` should trigger the events `moveCursorToStart(true)` and `moveCursorToEnd(true)` -- what they already were calling before?

4. Fixed redundant annotations

When annotations are no longer needed, IntelliJ would already notify us by making them grey. Always take the step to clean them up. In the case of this review, I `run inspection by name` and looked for `redundant annotation`, and removed them one by one as there were many. However, it turns out that IntelliJ was incorrectly identifying some annotation as "redundant". It turns out I had to specify the fully qualified import path to of the annotation class in order for IntelliJ to recognise them as being used in the class/function.

5. Privacy of Icon.Shape

Now that you've abstracted away the Icon.Shapes from being used directly in the application, by introducing `Icon` an enum that is defined by `Icon.Shape`, you should make `Icon.Shape` `private` so that no other module can use it other than `Icon`, as it is intended to only be used by `Icon`.

6. A lot of code formatting cleanup