Provide friendly warning when trying to explain an inference while explanations are not enabled (#689)
## What is the goal of this PR?

We have provided a friendly warning to users when they try to explain an
inference while explanations are not enabled.

Closes https://github.com/vaticle/typedb-studio/issues/608

## What are the changes implemented in this PR?

We have a new warning message popup, informing users that they must
enable explanations under the necessary circumstances.

To do so:
* We've changed the way we access a transaction within `GraphArea` and
`GraphBuilder`, removing it as a property from `GraphArea` and making
all changes to the graph through `GraphBuilder`.
* We've added a mechanism to `GraphBuilder` called
`transactionSnapshot`. This is a way to use the transaction that was
used to create the graph, but only if snapshot was enabled at the time
the graph was created and the current `transactionState.transaction` is
the same transaction as was used to create the graph.
* We have also changed the naming behind how `TransactionState` exposes
whether a configuration option like 'explain' is enabled for the given
transaction or not. Contrary to what one might assume,
`transactionState.explain.enabled` does not mean that explain is enabled
(which is what `transactionState.explain.value` does) - it means that
the explain option could be toggled through the UI. We've renamed
`enabled` to `canToggle` in this context.